Fix precedence of **
Precedence of ** was wrong on method parser.
This is because in most places where we were generating a parser for a list of operators, the list of operators was sometimes ordered in an unconvenient way (for example, * was before ** on that list).
In order to solve this a new parser was added that receives the operators list as a parameter, and returns a parser that will always try the greediest operators first (it orders them in a way that will have ** before *). Did this instead of just reordering the list of OPERATORS because the pattern ...OPERATORS.map(key) was getting repeated all over the place, but we could have also implemented it by reordering the OPERATORS constant.